Prefactor: Move scanner/mappers to pkg
