+++ pkg/scannerv4/enricher/fixedby/return.go
@@ -1,0 +1,4 @@
+package fixedby
+
+// Type is the type of data returned from the Enricher's Enrich method.
+const Type = "message/vnd.stackrox.scannerv4.fixedby; enricher=fixedby"
