+++ pkg/scannerv4/enricher/nvd/return.go
@@ -1,0 +1,4 @@
+package nvd
+
+// Type is the type of data returned from the Enricher's Enrich method.
+const Type = `message/vnd.stackrox.scannerv4.vulnerability; enricher=nvd schema=https://csrc.nist.gov/schema/nvd/api/2.0/source_api_json_2.0.schema`
