+++ pkg/scannerv4/mappers/mappers_test.go
