+++ pkg/scannerv4/updater/manual.go
@@ -1,0 +1,3 @@
+package updater
+
+const Name = "stackrox-manual"
