+++ scanner/enricher/nvd/nvd.go
@@ -21,6 +21,7 @@ import (
 	"github.com/quay/claircore"
 	"github.com/quay/claircore/libvuln/driver"
 	"github.com/quay/zlog"
+	"github.com/stackrox/rox/pkg/scannerv4/enricher/nvd"
 	"github.com/stackrox/rox/pkg/utils"
 	"github.com/stackrox/rox/pkg/uuid"
 )
@@ -33,13 +34,10 @@ var (
 )
 
 const (
-	// Type is the type of data returned from the Enricher's Enrich method.
-	Type = `message/vnd.stackrox.scannerv4.vulnerability; enricher=nvd schema=https://csrc.nist.gov/schema/nvd/api/2.0/source_api_json_2.0.schema`
-
 	// DefaultFeeds is the default place to look for CVE feeds.
 	DefaultFeeds = `https://services.nvd.nist.gov/rest/json/cves/2.0/`
 
-	// This appears above and must be the same.
+	// This appears in [nvd.Type] and must be the same.
 	name = `nvd`
 
 	// First year for the yearly CVE feeds: https://nvd.nist.gov/vuln/data-feeds
@@ -524,11 +522,11 @@ func (e *Enricher) Enrich(ctx context.Context, g driver.EnrichmentGetter, r *cla
 		}
 	}
 	if len(m) == 0 {
-		return Type, nil, nil
+		return nvd.Type, nil, nil
 	}
 	b, err := json.Marshal(m)
 	if err != nil {
-		return Type, nil, err
+		return nvd.Type, nil, err
 	}
-	return Type, []json.RawMessage{b}, nil
+	return nvd.Type, []json.RawMessage{b}, nil
 }
