+++ scanner/enricher/nvd/nvd_test.go
@@ -18,6 +18,7 @@ import (
 	"github.com/quay/claircore"
 	"github.com/quay/claircore/libvuln/driver"
 	"github.com/quay/zlog"
+	"github.com/stackrox/rox/pkg/scannerv4/enricher/nvd"
 )
 
 func TestConfigure(t *testing.T) {
@@ -314,7 +315,7 @@ func TestEnrich(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	if got, want := kind, Type; got != want {
+	if got, want := kind, nvd.Type; got != want {
 		t.Errorf("got: %q, want: %q", got, want)
 	}
 	want := map[string][]map[string]interface{}{
